progress
I've tried to combine the GMAC and nRF sotware modules and this was the progress:
After an unknown problen with Rev0, and corrupted file in Rev1 that stopped the compiling i made Rev0-1.
There is a weird situation where if i enable the Cache memory the GMAC module doesn't work but if i disable the Cache the SPI timings get messed up for the nRF module.
AFter a long trail-and-error progress i succeded in combining the 2 modules and receive Art-Net. now i have to make some prototype slave modules