Set-up for the distance-Hue connection
Did some tests for the Distance Hue connection, got it working.
When combined with the nRF base program the code executes 1 or 2 times and blocks.
I've got 2 options; 1) the problem lies in the network. 2) the Arduino is broken.
i'm looking into the RF24Mash library (an axtention of the RF24 library i'm already using to build a netwerk